Merge pull request #104 from FS-FAST-TRACK/feature/quizmaster-api-quizsession

+++ WebApp/backend/QuizMaster.API.Gatewway/Hubs/SessionHub.cs
@@ -1,9 +1,13 @@
 ﻿using Grpc.Net.Client;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
 using QuizMaster.API.Account.Proto;
+using QuizMaster.API.Authentication.Proto;
 using QuizMaster.API.Gateway.Configuration;
+using QuizMaster.API.Gateway.Helper;
 using QuizMaster.API.Gateway.Services;
 using QuizMaster.API.QuizSession.Models;
 using QuizMaster.API.QuizSession.Protos;
@@ -15,12 +19,13 @@ using System.Threading.Channels;
 
 namespace QuizMaster.API.Gateway.Hubs
 {
+
     public class SessionHub : Hub
     {
         private GrpcChannel _channel;
         private  QuizRoomService.QuizRoomServiceClient _channelClient;
+        private readonly AuthService.AuthServiceClient _authChannelClient;
         private SessionHandler SessionHandler;
-        private List<string> NAMES = new List<string>() { "Harold", "Jay", "JM", "Ada", "Pia"," Bo", "Rodney", "Neal", "Jess", "Aly", "James", "Xerxes", "Wayne", "Ken"};
         /*
          * TODO
          * - Create A QB Context for QuizParticipants
@@ -32,12 +37,30 @@ namespace QuizMaster.API.Gateway.Hubs
         public SessionHub(IOptions<GrpcServerConfiguration> options, SessionHandler sessionHandler)
         {
             _channel = GrpcChannel.ForAddress(options.Value.Session_Service);
-            _channelClient = new QuizRoomService.QuizRoomServiceClient(_channel); 
+            _channelClient = new QuizRoomService.QuizRoomServiceClient(_channel);
+            _channel = GrpcChannel.ForAddress(options.Value.Authentication_Service);
+            _authChannelClient = new AuthService.AuthServiceClient(_channel);
             SessionHandler = sessionHandler;
+        }
+
+        /*
+         * Login the User
+         */
+        public async Task Login(string token)
+        {
+            var connectionId = Context.ConnectionId;
+            await SessionHandler.AuthenticateConnectionId(this, _authChannelClient, connectionId, token);
         }
 
         public async Task CreateRoom(CreateRoomDTO roomDTO)
         {
+            string connectionId = Context.ConnectionId;
+            if (!SessionHandler.IsAdmin(connectionId))
+            {
+                await Clients.Caller.SendAsync("notif", "Please contact administrator");
+                return;
+            }
+
             var request = new CreateRoomRequest { Room = JsonConvert.SerializeObject(roomDTO) };
             var reply = await _channelClient.CreateRoomAsync(request);
 
@@ -58,6 +81,13 @@ namespace QuizMaster.API.Gateway.Hubs
 
         public async Task DeleteRoom(int roomId)
         {
+            string connectionId = Context.ConnectionId;
+            if (!SessionHandler.IsAdmin(connectionId))
+            {
+                await Clients.Caller.SendAsync("notif", "Please contact administrator");
+                return;
+            }
+
             var request = new ModifyRoomRequest { Room = roomId };
 
             var reply = await _channelClient.DeleteRoomAsync(request);
@@ -76,6 +106,12 @@ namespace QuizMaster.API.Gateway.Hubs
 
         public async Task UpdateRoom(UpdateRoomDTO updateRoomDTO)
         {
+            string connectionId = Context.ConnectionId;
+            if (!SessionHandler.IsAdmin(connectionId))
+            {
+                await Clients.Caller.SendAsync("notif", "Please contact administrator");
+                return;
+            }
             var request = new CreateRoomRequest { Room = JsonConvert.SerializeObject(updateRoomDTO) };
             var reply = await _channelClient.UpdateRoomAsync(request);
 
@@ -97,14 +133,35 @@ namespace QuizMaster.API.Gateway.Hubs
         {
             string connectionId = Context.ConnectionId;
 
+            if (!SessionHandler.IsAuthenticated(connectionId))
+            {
+                await Clients.Caller.SendAsync("chat", "You need to login to join chat");
+                return;
+            }
+            var participantData = SessionHandler.GetLinkedParticipantInConnectionId(connectionId);
+            if(participantData == null) { return; }
             // send chat only to group
-            await Clients.Group(roomId).SendAsync("chat", $"[{SessionHandler.GetLinkedParticipantInConnectionId(connectionId).QParticipantDesc}]: {chat}");
+            if (SessionHandler.IsAdmin(connectionId))
+                await Clients.Group(roomId).SendAsync("chat", $"[{participantData.QParticipantDesc}(host)]: {chat}");
+            else await Clients.Group(roomId).SendAsync("chat", $"[{participantData.QParticipantDesc}]: {chat}");
         }
-
 
         public async Task JoinRoom(int RoomPin)
         {
             string connectionId = Context.ConnectionId;
+            if (!SessionHandler.IsAuthenticated(connectionId))
+            {
+                await Clients.Caller.SendAsync("notif", "Failed to join room, you are not authorized");
+                return;
+            }
+            // Grab the UserInformation
+            var userData = await SessionHandler.GetUserInformation(_authChannelClient, SessionHandler.GetConnectionToken(connectionId));
+            if(userData == null)
+            {
+                await Clients.Caller.SendAsync("notif", $"There was an issue authenticating your account");
+                return;
+            }
+
             try
             {
                 var reply = _channelClient.GetAllRoom(new RoomsEmptyRequest());
@@ -127,16 +184,18 @@ namespace QuizMaster.API.Gateway.Hubs
 
                     if (containsId)
                     {
-                        string Name = NAMES[new Random().Next(0, NAMES.Count - 1)];
-                        SessionHandler.LinkParticipantConnectionId(connectionId, new QuizParticipant { QParticipantDesc = Name });
+                        //string Name = NAMES[new Random().Next(0, NAMES.Count - 1)];
+                        string Name = userData.UserData.UserName;
+                        SessionHandler.LinkParticipantConnectionId(connectionId, new QuizParticipant { QParticipantDesc = Name, UserId = userData.UserData.Id });
                         await SessionHandler.AddToGroup(this, $"{RoomPin}", connectionId);
-                        await Clients.Group($"{RoomPin}").SendAsync("notif",$"{Name} has joined Room {room.QRoomDesc}", room );
+                        await Clients.Group($"{RoomPin}").SendAsync("notif", $"{Name} has joined Room {room.QRoomDesc}", room);
                     }
                     //await Clients.All.SendAsync("QuizRooms", quizRooms);
                 }
         }
             catch (Exception ex)
             {
+                await Clients.Caller.SendAsync("notif", $"An error has occurred while trying to connect");
                 Console.Write(ex.ToString());
             }
         }
@@ -162,16 +221,25 @@ namespace QuizMaster.API.Gateway.Hubs
         }
         public override async Task OnDisconnectedAsync(Exception? exception)
         {
-            await SessionHandler.RemoveClientFromGroups(this, Context.ConnectionId, $"{SessionHandler.GetLinkedParticipantInConnectionId(Context.ConnectionId).QParticipantDesc} has been disconnected");
+            await LeaveRoom();
         }
 
         public async Task LeaveRoom()
         {
-            await SessionHandler.RemoveClientFromGroups(this, Context.ConnectionId, $"{SessionHandler.GetLinkedParticipantInConnectionId(Context.ConnectionId).QParticipantDesc} has left the room");
+            var participantData = SessionHandler.GetLinkedParticipantInConnectionId(Context.ConnectionId);
+            if (participantData == null) { return; }
+            await SessionHandler.RemoveClientFromGroups(this, Context.ConnectionId, $"{participantData.QParticipantDesc} has left the room");
+            SessionHandler.UnbindConnectionId(Context.ConnectionId);
         }
 
         public async Task StartRoom(string roomPin)
         {
+            string connectionId = Context.ConnectionId;
+            if (!SessionHandler.IsAdmin(connectionId))
+            {
+                await Clients.Caller.SendAsync("notif", "Please contact administrator");
+                return;
+            }
             try
             {
                 var reply = _channelClient.GetAllRoom(new RoomsEmptyRequest());
@@ -205,56 +273,6 @@ namespace QuizMaster.API.Gateway.Hubs
             catch (Exception ex)
             {
                 Console.Write(ex?.ToString());
-            }
-        }
-
-        public async Task GetQuestionSets(int id)
-        {
-            var set = new SetRequest() { Id = id };
-            var setReply = await _channelClient.GetQuizSetAsync(set);
-
-            if(setReply.Code == 200)
-            {
-                var quizSets = JsonConvert.DeserializeObject<List<SetQuizRoom>>(setReply.Data);
-
-                var roomPin = SessionHandler.GetConnectionGroup(Context.ConnectionId);
-                if (roomPin != null)
-                {
-                    await Clients.Group(roomPin).SendAsync("questionSet", quizSets);
-                }
-            }
-        }
-
-        public async Task GetQuestions(int id)
-        {
-            var request = new SetRequest() { Id = id };
-            var reply = await _channelClient.GetQuizAsync(request);
-
-            if (reply.Code == 200)
-            {
-                var questions = JsonConvert.DeserializeObject<List<QuestionSet>>(reply.Data);
-                var roomPin = SessionHandler.GetConnectionGroup(Context.ConnectionId);
-                if (roomPin != null)
-                {
-                    foreach(var question in questions)
-                    {
-                        var questionRequest = new SetRequest() { Id = question.QuestionId };
-                        var questionReply = _channelClient.GetQuestion(questionRequest);
-
-                        if(questionReply.Code == 200) 
-                        {
-                            var details = JsonConvert.DeserializeObject<QuestionsDTO>(questionReply.Data);
-                            var timout = details.question.QTime;
-
-                            for(int time = timout; time > 0; time--)
-                            {
-                                details.RemainingTime = time;
-                                await Clients.Group(roomPin).SendAsync("question", details);
-                                await Task.Delay(1000);
-                            }
-                        }
-                    }
-                }
             }
         }
